Added YusufUser utility
