+++ jda5/src/main/java/io/github/yusufsdiscordbot/yusufsdiscordcore/bot/slash_command/core/YusufUser.java
@@ -13,6 +13,7 @@
 
 package io.github.yusufsdiscordbot.yusufsdiscordcore.bot.slash_command.core;
 
+import io.github.yusufsdiscordbot.yusufsdiscordcore.bot.slash_command.core.utility.YusufUserUtility;
 import io.github.yusufsdiscordbot.yusufsdiscordcore.bot.slash_command.interactions.YusufSlashCommandEvent;
 import net.dv8tion.jda.api.JDA;
 import net.dv8tion.jda.api.entities.Guild;
@@ -27,7 +28,13 @@ import java.util.List;
 import java.util.Set;
 
 @SuppressWarnings("unused")
-public record YusufUser(User user) implements IMentionable {
+public class YusufUser extends YusufUserUtility implements IMentionable {
+    private final User user;
+
+    public YusufUser(User user) {
+        super(user);
+        this.user = user;
+    }
 
     /**
      * @see User
@@ -160,100 +167,6 @@ public record YusufUser(User user) implements IMentionable {
      */
     public @NotNull String getAsMention() {
         return this.user.getAsMention();
-    }
-
-    /**
-     * @see YusufBot
-     */
-    public @NotNull YusufBot getBot() {
-        return new YusufBot(this.user.getJDA().getSelfUser());
-    }
-
-    /**
-     * @see YusufBot#getBotTag()
-     */
-    @Contract(pure = true)
-    public @NotNull String getBotTag() {
-        return this.getBot().getBotTag();
-    }
-
-    /**
-     * @see YusufBot#getBotId()
-     */
-    @Contract(pure = true)
-    public @NotNull String getBotId() {
-        return this.getBot().getBotId();
-    }
-
-    /**
-     * @see YusufBot#getBotIdLong()
-     */
-    @Contract(pure = true)
-    public @NotNull Long getBotIdLong() {
-        return this.getBot().getBotIdLong();
-    }
-
-    /**
-     * @see YusufBot#getName()
-     */
-    @Contract(pure = true)
-    public @NotNull String getBotName() {
-        return this.getBot().getName();
-    }
-
-    /**
-     * @see YusufBot#getDiscriminator()
-     */
-    @Contract(pure = true)
-    public @NotNull String getBotDiscriminator() {
-        return this.getBot().getDiscriminator();
-    }
-
-    /**
-     * @see YusufBot#getAvatarId()
-     */
-    @Contract(pure = true)
-    public @NotNull String getBotAvatarId() {
-        return this.getBot().getAvatarId();
-    }
-
-    /**
-     * @see YusufBot#getAvatarUrl()
-     */
-    @Contract(pure = true)
-    public @NotNull String getBotAvatarUrl() {
-        return this.getBot().getAvatarUrl();
-    }
-
-    /**
-     * @see YusufBot#getDefaultAvatarId()
-     */
-    @Contract(pure = true)
-    public @NotNull String getBotDefaultAvatarId() {
-        return this.getBot().getDefaultAvatarId();
-    }
-
-    /**
-     * @see YusufBot#getBotEffectiveAvatarUrl()
-     */
-    @Contract(pure = true)
-    public @NotNull String getBotEffectiveAvatarUrl() {
-        return this.getBot().getBotEffectiveAvatarUrl();
-    }
-
-    /**
-     * @param user the user who you want to check is null
-     * @param event the slash command event.
-     * @return not null
-     */
-    @Contract("null,_->false;!null,_->true")
-    public boolean userIsNotNull(YusufUser user, YusufSlashCommandEvent event) {
-        boolean result = user == null;
-        if (result) {
-            event.replyEphemeral("The given user is null");
-            return false;
-        }
-        return true;
     }
 
     @Override
